fix import architecture
