+++ libs/shared/constants/README.md
@@ -1,0 +1,11 @@
+# shared-constants
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test shared-constants` to execute the unit tests via [Jest](https://jestjs.io).
+
+## Running lint
+
+Run `nx lint shared-constants` to execute the lint via [ESLint](https://eslint.org/).
