+++ libs/shared/constants/src/lib/profile-picture.ts
@@ -1,0 +1,13 @@
+import {
+  ValidProfilePictureFileExtension,
+  ValidProfilePictureMimeType,
+} from '@bregenz-bewegt/shared/types';
+
+export const validProfilePictureFileExtensions: ValidProfilePictureFileExtension[] =
+  ['jpeg', 'jpg', 'png', 'gif'];
+export const validProfilePictureMimeTypes: ValidProfilePictureMimeType[] = [
+  'image/jpeg',
+  'image/jpg',
+  'image/png',
+  'image/gif',
+];
