+++ libs/shared/types/src/lib/types/common/profile-picture.ts
@@ -1,12 +1,3 @@
 export type ValidProfilePictureFileExtension = 'jpeg' | 'jpg' | 'png' | 'gif';
 export type ValidProfilePictureMimeType =
   `image/${ValidProfilePictureFileExtension}`;
-
-export const validProfilePictureFileExtensions: ValidProfilePictureFileExtension[] =
-  ['jpeg', 'jpg', 'png', 'gif'];
-export const validProfilePictureMimeTypes: ValidProfilePictureMimeType[] = [
-  'image/jpeg',
-  'image/jpg',
-  'image/png',
-  'image/gif',
-];
