+++ tsconfig.base.json
@@ -60,6 +60,9 @@
       "@bregenz-bewegt/server/multer": ["libs/server/multer/src/index.ts"],
       "@bregenz-bewegt/server/types": ["libs/server/types/src/index.ts"],
       "@bregenz-bewegt/server/util": ["libs/server/util/src/index.ts"],
+      "@bregenz-bewegt/shared/constants": [
+        "libs/shared/constants/src/index.ts"
+      ],
       "@bregenz-bewegt/shared/errors": ["libs/shared/errors/src/index.ts"],
       "@bregenz-bewegt/shared/types": ["libs/shared/types/src/index.ts"]
     }
