+++ workspace.json
@@ -25,6 +25,7 @@
     "server-prisma": "libs/server/prisma",
     "server-types": "libs/server/types",
     "server-util": "libs/server/util",
+    "shared-constants": "libs/shared/constants",
     "shared-errors": "libs/shared/errors",
     "shared-types": "libs/shared/types"
   }
